feature: render different components (technician and householder) depending on condition in same view to increase reusability
